Make location auto-detection configurable.
This is a rather large change in that it removes all arguments to
`fetch-forecast.py` from `index.coffee`. Config items are things like
API keys and binary locations. Options and arguments to the script are
things like location auto-detection, reverse geocoding, and providing a
latitude and longitude or a location. These arguments are located in the
fetch-forecast.args file, one per line.